sageai-sample/requirements.txt
python-dotenv==1.1.1
langchain==0.3.27
langchain_community==0.3.29
llama-index==0.14.2
pinecone==7.3.0
llama-index-vector-stores-pinecone==0.7.1
fastapi==0.116.1
uvicorn==0.35.0
markdown-it-py[plugins,linkify]==4.0.0
linkify-it-py==2.0.3
mdit-py-plugins==0.5.0